refactor(server): remove barrel exports in favor of direct imports
Delete 10 barrel index.ts files across common/, shared/ and modules/*/entities,
and update all consumers to import from concrete source files. Aligns with the
project's no-barrel-export convention and avoids unnecessary re-export layers.
