+++ apps/server/src/app.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get } from '@nestjs/common';
 import { AppService } from './app.service';
-import { Public } from './common/decorators';
+import { Public } from './common/decorators/jwt-auth.decorator';
 
 @Controller()
 export class AppController {
