--- apps/server/src/common/constants/index.ts
@@ -1,3 +1,0 @@
-export * from './environment';
-export * from './permissions';
-export * from './roles';
