--- apps/server/src/common/context/index.ts
@@ -1,1 +1,0 @@
-export * from './user-context';
