--- apps/server/src/common/decorators/index.ts
@@ -1,5 +1,0 @@
-export * from './disabled-endpoint.decorator';
-export * from './group-member-roles.decorator';
-export * from './jwt-auth.decorator';
-export * from './permission.decorator';
-export * from './special-roles.decorator';
