--- apps/server/src/common/guards/index.ts
@@ -1,4 +1,0 @@
-export * from './disabled-endpoint.guard';
-export * from './jwt-auth.guard';
-export * from './permission.guard';
-export * from './special-roles.guard';
