--- apps/server/src/modules/cats/entities/index.ts
@@ -1,1 +1,0 @@
-export * from './cat.entity';
