--- apps/server/src/modules/groups/entities/index.ts
@@ -1,2 +1,0 @@
-export * from './group-member.entity';
-export * from './group.entity';
