+++ apps/server/src/modules/groups/groups.module.ts
@@ -1,7 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UsersModule } from '../users/users.module';
-import { Group, GroupMember } from './entities';
+import { GroupMember } from './entities/group-member.entity';
+import { Group } from './entities/group.entity';
 import { GroupsController } from './groups.controller';
 import { GroupsService } from './groups.service';
