--- apps/server/src/modules/permissions/entities/index.ts
@@ -1,1 +1,0 @@
-export * from './permission.entity';
