--- apps/server/src/modules/roles/entities/index.ts
@@ -1,1 +1,0 @@
-export * from './role.entity';
