+++ apps/server/src/modules/roles/roles.controller.ts
@@ -1,4 +1,4 @@
-import { PermissionCode } from '@/common/constants';
+import { PermissionCode } from '@/common/constants/permissions';
 import { Permission } from '@/common/decorators/permission.decorator';
 import {
   SpecialRoles,
