--- apps/server/src/modules/users/entities/index.ts
@@ -1,1 +1,0 @@
-export * from './user.entity';
