--- apps/server/src/shared/utils/index.ts
@@ -1,4 +1,0 @@
-export * from './extract-token';
-export * from './roles';
-export * from './snowflake';
-export * from './swagger';
